+++ CHANGELOG.md
@@ -61,6 +61,8 @@ The format is based on [Keep a Changelog] and this project adheres to
 - `dispatch` module for [Dispatch](https://developer.apple.com/documentation/dispatch)
   library.
 
+  - Types: `DispatchObject`.
+
 - Pointer methods for `SEL`.
 
 - Implemented `PartialEq` for `NSObject`.